@@ -16,27 +16,27 @@
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.1.0",
-    "@docusaurus/preset-classic": "^2.1.0",
-    "@heroicons/react": "^2.0.10",
-    "@mdx-js/react": "^1.6.22",
-    "clsx": "^1.2.1",
-    "hast-util-is-element": "^1.1.0",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-player": "^2.11.0",
-    "redocusaurus": "^1.4.0",
-    "rehype-katex": "^5.0.0",
-    "remark-math": "^3.0.1"
+    "@docusaurus/core": "^3.0.0",
+    "@docusaurus/preset-classic": "^3.0.0",
+    "@heroicons/react": "^2.0.18",
+    "@mdx-js/react": "^3.0.0",
+    "clsx": "^2.0.0",
+    "hast-util-is-element": "^3.0.0",
+    "prism-react-renderer": "^2.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-player": "^2.13.0",
+    "redocusaurus": "^2.0.0",
+    "rehype-katex": "^7.0.0",
+    "remark-math": "^6.0.0"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.1.0",
-    "@tsconfig/docusaurus": "^1.0.6",
-    "autoprefixer": "^10.4.8",
+    "@docusaurus/module-type-aliases": "^3.0.0",
+    "@tsconfig/docusaurus": "^2.0.2",
+    "autoprefixer": "^10.4.16",
     "postcss": "^8.4.31",
-    "tailwindcss": "^3.1.8",
-    "typescript": "^4.7.4"
+    "tailwindcss": "^3.3.5",
+    "typescript": "^5.2.2"
   },
   "browserslist": {
     "production": [
